💸 Zoopla sale adverts: skip saving duplicate content in S3.
Saves not only unnecessary S3 storage, but more importantly S3 API calls.

Writes to a new DynamoDB table as an "index table" instead, but that
still works out cheaper than S3 - both in terms of storage and requests.

This is feasible because the content size is only ~350KB (could be
gzipped to ~50KB if the need arises in future).